+++ Rakefile
@@ -18,8 +18,8 @@ task :dependapanda_loud do
 end
 
 task :save_application_gemfiles do
-  UseCases::SaveGemfiles.new(
-    fetch_gemfiles: UseCases::FetchGemfiles.new(
+  UseCases::Gemfiles::Save.new(
+    fetch_gemfiles: UseCases::Gemfiles::Fetch.new(
       teams_use_case: UseCases::FetchTeams.new
     )
   ).execute
