+++ lib/loader.rb
@@ -13,8 +13,9 @@ require_relative './use_cases/group_pull_requests_by_gem'
 require_relative './use_cases/slack/schedulers/every_day'
 require_relative './use_cases/slack/schedulers/weekday'
 require_relative './use_cases/slack/send_messages'
-require_relative './use_cases/fetch_gemfiles'
-require_relative './use_cases/save_gemfiles'
+require_relative './use_cases/gemfiles/fetch'
+require_relative './use_cases/gemfiles/save'
+
 require_relative './gateways/pull_request'
 require_relative './gateways/pull_request_count'
 require_relative './gateways/slack_message'
